peers/requirements.txt
Django
djangorestframework
drfdocs
django-celery
redis
